Change chainId env to network name (#17)
* Change chainId env to network name

* Update src/services/signer.ts

Co-authored-by: qdqd <qdqdqdqdqd@protonmail.com>

* 🩹 manage the case when process.env.PORT is empty

As environment variables are not typed, if the PORT environment variable
is present in the .env file but not set (PORT=), the process.env.PORT will
be an empty string not an undefined. As this is not a valid port number,
so we need to handle this case. To manage this I switched from the nullish
coalescing operator to the logical OR operator in the condition.

* 🩹 fix starknet import in the script

Import reorg introduce by #96976d2 was not correct, this commit fixes it.

---------

Co-authored-by: qdqd <qdqdqdqdqd@protonmail.com>