✅ "improve" RPCProvider mocking
Using a custom class implementation that extends the RPCProvider class,
we now have more control on the mocking of getEvents and a way to intercept
all http requests made by the provider. I would say "improve" is a bit exagerated
in this case but this implementation is more flexible and give us the foundation
to mock any methods by tests.